Refactor update process: Split DownloadInstallPageViewModel into DownloadPageViewModel and InstallPageViewModel
- Removed DownloadInstallPageViewModel and its associated XAML and code-behind files.
- Created DownloadPageViewModel to handle the download process separately.
- Created InstallPageViewModel to manage the installation process.
- Updated App.xaml.cs to register new view models.
- Adjusted navigation logic in UpdateCheckPageViewModel to navigate to DownloadPage instead of DownloadInstallPage.
- Updated XAML and code-behind for DownloadPage and InstallPage to reflect new structure and functionality.
