+++ src/Bucket.Updater/Views/DownloadPage.xaml
@@ -1,0 +1,80 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Page x:Class="Bucket.Updater.Views.DownloadPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:dev="using:DevWinUI"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      mc:Ignorable="d">
+
+    <Grid Padding="32">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="*" />
+        </Grid.RowDefinitions>
+
+        <!-- Header -->
+        <StackPanel Grid.Row="0" Spacing="8" HorizontalAlignment="Center" Margin="0,0,0,32">
+            <TextBlock Text="{x:Bind ViewModel.HeaderText, Mode=OneWay}" 
+                       FontSize="28" 
+                       FontWeight="Bold"
+                       HorizontalAlignment="Center" />
+            <TextBlock Text="{x:Bind ViewModel.UpdateVersion, Mode=OneWay}"
+                       FontSize="16"
+                       Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                       HorizontalAlignment="Center" />
+        </StackPanel>
+
+        <!-- Download Content (Centered) -->
+        <StackPanel Grid.Row="1" 
+                    VerticalAlignment="Center"
+                    HorizontalAlignment="Center"
+                    Spacing="32"
+                    MaxWidth="500">
+            
+            <!-- Progress Ring -->
+            <dev:ProgressRing IsActive="{x:Bind ViewModel.IsProgressActive, Mode=OneWay}"
+                              Width="64" 
+                              Height="64"
+                              HorizontalAlignment="Center" />
+
+            <!-- Progress Bar -->
+            <ProgressBar Value="{x:Bind ViewModel.ProgressPercentage, Mode=OneWay}"
+                         Minimum="0"
+                         Maximum="100"
+                         HorizontalAlignment="Stretch" />
+
+            <!-- Status Text -->
+            <StackPanel Spacing="8" 
+                        HorizontalAlignment="Center">
+                <TextBlock Text="{x:Bind ViewModel.StatusMessage, Mode=OneWay}" 
+                           FontSize="18"
+                           FontWeight="SemiBold"
+                           HorizontalAlignment="Center"
+                           TextAlignment="Center" />
+                
+                <!-- Download specific info -->
+                <StackPanel Orientation="Horizontal" 
+                            Spacing="8"
+                            HorizontalAlignment="Center">
+                    <TextBlock Text="{x:Bind ViewModel.DownloadedSize, Mode=OneWay}" />
+                    <TextBlock Text="/" />
+                    <TextBlock Text="{x:Bind ViewModel.TotalSize, Mode=OneWay}" />
+                </StackPanel>
+                
+                <TextBlock Text="{x:Bind ViewModel.DownloadSpeed, Mode=OneWay}"
+                           Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                           HorizontalAlignment="Center" />
+            </StackPanel>
+
+            <!-- Cancel Button (only during download) -->
+            <Button x:Name="CancelButton"
+                    Content="Cancel"
+                    Command="{x:Bind ViewModel.CancelCommand}"
+                    HorizontalAlignment="Center"
+                    Visibility="{x:Bind ViewModel.CancelButtonVisibility, Mode=OneWay}" />
+
+        </StackPanel>
+
+    </Grid>
+</Page>
