+++ src/Bucket.Updater/Views/DownloadPage.xaml.cs
@@ -1,0 +1,32 @@
+namespace Bucket.Updater.Views
+{
+    public sealed partial class DownloadPage : Page
+    {
+        public DownloadPageViewModel ViewModel { get; }
+
+        public DownloadPage()
+        {
+            this.InitializeComponent();
+            ViewModel = App.GetService<DownloadPageViewModel>();
+            DataContext = ViewModel;
+        }
+
+        protected override void OnNavigatedTo(NavigationEventArgs e)
+        {
+            base.OnNavigatedTo(e);
+            
+            if (e.Parameter is Bucket.Updater.Models.UpdateInfo updateInfo)
+            {
+                ViewModel.StartDownload(updateInfo);
+            }
+        }
+
+        protected override void OnNavigatedFrom(NavigationEventArgs e)
+        {
+            base.OnNavigatedFrom(e);
+            
+            // Cleanup any resources if needed
+            ViewModel.Cleanup();
+        }
+    }
+}
